Move mount package into libcontainer root
