+++ linux_userns_sidecar_init.go
@@ -5,7 +5,6 @@ package libcontainer
 import (
 	"github.com/docker/libcontainer/configs"
 	"github.com/docker/libcontainer/label"
-	"github.com/docker/libcontainer/mount"
 )
 
 // linuxUsernsSideCar is run to setup mounts and networking related operations
@@ -28,7 +27,7 @@ func (l *linuxUsernsSideCar) Init() error {
 	label.Init()
 	// InitializeMountNamespace() can be executed only for a new mount namespace
 	if l.config.Config.Namespaces.Contains(configs.NEWNET) {
-		if err := mount.InitializeMountNamespace(l.config.Config); err != nil {
+		if err := setupRootfs(l.config.Config); err != nil {
 			return err
 		}
 	}
