--- mount/readonly.go
@@ -1,11 +1,0 @@
-// +build linux
-
-package mount
-
-import (
-	"syscall"
-)
-
-func setReadonly() error {
-	return syscall.Mount("/", "/", "bind", syscall.MS_BIND|syscall.MS_REMOUNT|syscall.MS_RDONLY|syscall.MS_REC, "")
-}
